+++ src/main/kotlin/net/pwall/json/JSONDeserializer.kt
@@ -54,7 +54,6 @@ import java.util.Date
 import java.util.LinkedList
 import java.util.UUID
 
-import net.pwall.json.annotation.JSONName
 import net.pwall.util.ISO8601Date
 
 /**
@@ -72,8 +71,8 @@ object JSONDeserializer {
      * @param   config      an optional [JSONConfig]
      * @return              the converted object
      */
-    fun deserialize(resultType: KType, json: JSONValue?, config: JSONConfig? = null): Any? {
-        config?.getFromJSONMapping(resultType)?.let { return it(json) }
+    fun deserialize(resultType: KType, json: JSONValue?, config: JSONConfig = JSONConfig.defaultConfig): Any? {
+        config.getFromJSONMapping(resultType)?.let { return it(json) }
         if (json == null) {
             if (!resultType.isMarkedNullable)
                 throw JSONException("Can't deserialize null as $resultType")
@@ -90,7 +89,8 @@ object JSONDeserializer {
      * @param   json        the parsed JSON, as a [JSONValue] (or `null`)
      * @return              the converted object
      */
-    fun <T: Any> deserialize(resultClass: KClass<T>, json: JSONValue?, config: JSONConfig? = null): T? {
+    fun <T: Any> deserialize(resultClass: KClass<T>, json: JSONValue?,
+            config: JSONConfig = JSONConfig.defaultConfig): T? {
         if (json == null)
             return null
         return deserialize(resultClass, emptyList(), json, config)
@@ -103,7 +103,8 @@ object JSONDeserializer {
      * @param   json        the parsed JSON, as a [JSONValue] (or `null`)
      * @return              the converted object
      */
-    fun <T: Any> deserializeNonNull(resultClass: KClass<T>, json: JSONValue?, config: JSONConfig? = null): T {
+    fun <T: Any> deserializeNonNull(resultClass: KClass<T>, json: JSONValue?,
+            config: JSONConfig = JSONConfig.defaultConfig): T {
         if (json == null)
             throw JSONException("Can't deserialize null as ${resultClass.simpleName}")
         return deserialize(resultClass, emptyList(), json, config)
@@ -120,7 +121,7 @@ object JSONDeserializer {
      */
     @Suppress("UNCHECKED_CAST")
     fun <T: Any> deserialize(resultClass: KClass<T>, types: List<KTypeProjection>, json: JSONValue,
-            config: JSONConfig? = null): T {
+            config: JSONConfig = JSONConfig.defaultConfig): T {
 
         // check for JSONValue
 
@@ -260,7 +261,7 @@ object JSONDeserializer {
 
     @Suppress("UNCHECKED_CAST", "IMPLICIT_CAST_TO_ANY")
     fun <T: Any> deserializeArray(resultClass: KClass<T>, types: List<KTypeProjection>, json: JSONArray,
-            config: JSONConfig?): T {
+            config: JSONConfig): T {
         try {
 
             return when (resultClass) {
@@ -370,7 +371,7 @@ object JSONDeserializer {
 
     @Suppress("UNCHECKED_CAST")
     fun <T: Any> deserializeObject(resultClass: KClass<T>, types: List<KTypeProjection>, json: JSONObject,
-            config: JSONConfig?): T {
+            config: JSONConfig): T {
 
         try {
             if (resultClass.isSubclassOf(Map::class)) {
@@ -388,11 +389,11 @@ object JSONDeserializer {
 
             resultClass.objectInstance?.let { return setRemainingFields(resultClass, it, json, config) }
 
-            findBestConstructor(resultClass.constructors, json)?.let { constructor ->
+            findBestConstructor(resultClass.constructors, json, config)?.let { constructor ->
                 val argMap = HashMap<KParameter, Any?>()
                 val jsonCopy = HashMap<String, JSONValue?>(json)
                 constructor.parameters.forEach { parameter ->
-                    val paramName = parameter.aName()
+                    val paramName = findParameterName(parameter, config)
                     jsonCopy[paramName]?.let {
                         argMap[parameter] = deserialize(parameter.type, it, config)
                         jsonCopy.remove(paramName)
@@ -413,9 +414,9 @@ object JSONDeserializer {
     }
 
     private fun <T: Any>  setRemainingFields(resultClass: KClass<T>, instance: T, json: Map<String, JSONValue?>,
-            config: JSONConfig?): T {
+            config: JSONConfig): T {
         json.forEach { entry -> // JSONObject fields not used in constructor
-            val member = findField(resultClass.members, entry.key) ?:
+            val member = findField(resultClass.members, entry.key, config) ?:
                     throw JSONException("Can't find property ${entry.key} in ${resultClass.simpleName}")
             val value = deserialize(member.returnType, json[entry.key], config)
             if (member is KMutableProperty<*>) {
@@ -439,22 +440,23 @@ object JSONDeserializer {
         return instance
     }
 
-    private fun findField(members: Collection<KCallable<*>>, name: String): KProperty<*>? {
+    private fun findField(members: Collection<KCallable<*>>, name: String, config: JSONConfig): KProperty<*>? {
         for (member in members) {
-            if (member is KProperty<*> && member.aName() == name)
+            if (member is KProperty<*> && (config.findNameFromAnnotation(member.annotations) ?: member.name) == name)
                     return member
         }
         return null
     }
 
-    private fun <T: Any> findBestConstructor(constructors: Collection<KFunction<T>>, json: JSONObject): KFunction<T>? {
+    private fun <T: Any> findBestConstructor(constructors: Collection<KFunction<T>>, json: JSONObject,
+            config: JSONConfig): KFunction<T>? {
         var result: KFunction<T>? = null
         var best = -1
         for (constructor in constructors) {
             val parameters = constructor.parameters
-            if (parameters.any { it.aName() == null || it.kind != KParameter.Kind.VALUE })
+            if (parameters.any { findParameterName(it, config) == null || it.kind != KParameter.Kind.VALUE })
                 continue
-            val n = findMatchingParameters(parameters, json)
+            val n = findMatchingParameters(parameters, json, config)
             if (n > best) {
                 result = constructor
                 best = n
@@ -463,10 +465,10 @@ object JSONDeserializer {
         return result
     }
 
-    private fun findMatchingParameters(parameters: List<KParameter>, json: JSONObject): Int {
+    private fun findMatchingParameters(parameters: List<KParameter>, json: JSONObject, config: JSONConfig): Int {
         var n = 0
         for (parameter in parameters) {
-            if (json.containsKey(parameter.aName()))
+            if (json.containsKey(findParameterName(parameter, config)))
                 n++
             else {
                 if (!parameter.isOptional)
@@ -476,13 +478,8 @@ object JSONDeserializer {
         return n
     }
 
-    private fun KParameter.aName(): String? { // annotated name, or regular name if not overridden
-        return findAnnotation<JSONName>()?.name ?: name
-    }
-
-    private fun KProperty<*>.aName(): String? { // annotated name, or regular name if not overridden
-        return findAnnotation<JSONName>()?.name ?: name
-    }
+    private fun findParameterName(parameter: KParameter, config: JSONConfig): String? =
+            config.findNameFromAnnotation(parameter.annotations) ?: parameter.name
 
     private val fromJsonCache = HashMap<KClass<*>, KFunction<*>>()